Change conveyor behavior
